fix: export DEFAULT_PREFERENCES from kolu-common/config
Importing from the barrel index.ts forced vite to bundle all of
kolu-common, transitively pulling in claude-agent-sdk and breaking
the client build. The config entrypoint is already used by the
client for other constants and doesn't reach server-only deps.
